Change how bind mount sources are treated
BREAKING CHANGE:
Require a full path in group_vars/fileserver.yaml, rather than assuming
this will happen within a certain drive context.

The reason for this is because the context was initially added due to a
specialised set up with my cluster. Regardless, this is not a good
assumption to make and reduces the versatility of these binds.