Write protect mount points if drives find themselves unmounted
New task to perform `chattr +i` on mountpoint directories. This makes
the directories immutable if no mounted drive.

This includes a clean up on how default values are set up as the mount
point path is needed before the drive is mounted to perform this
operation. As a bonus, we are now able to remove indexed loops and
simplify certain tasks.